requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.0
av==11.0.0
numpy==1.26.0
python-dotenv==1.0.0
python-multipart
